Recording
The webpage shows two buttons (start and stop) for recording. When the
user clicks on stop recording button, the recording stops and the data
is saved on the server (./uploads directory). The 'workerid' is used as
the file name of the audio file. This file path along with workerid and
other parameters, is stored in the Mongodb database.